Revert "Changed Volume's datatype from int to long"
This reverts commit a7a3e5bba63dd5186aec94faf80749e1c46d8d09.

+++ ASX.BusinessLayer/EndOfDay.cs
@@ -25,7 +25,7 @@ namespace ASX.BusinessLayer
         public virtual decimal High { get; set; }
         public virtual decimal Low { get; set; }
         public virtual decimal Close { get; set; }
-        public virtual long Volume { get; set; }
+        public virtual int Volume { get; set; }
 
         public virtual Company Company { get; set; }
